Added file locking
I have added a file lock to prevent multiple instances of Tomaten Tijd
to run at the same time. I've used the article 'Elegant locking of a Bash program'
by Kfir Lavi[1] to do this while at the same time rewrote the code inspired
by the same article. Tomaten Tijd now will also tell you if there is
already an instance running both in /var/log/syslog and in the terminal
using logger.

The only thing left is having an indicator to
display the app is running. Feel free to send my any info on how to
achieve this with Bash.

[1] http://www.kfirlavi.com/blog/2012/11/06/elegant-locking-of-bash-program/